[5.2] Implement (some) expressions.
Since we're not writing Java, the utility of the expression-generator is
limited (we barely need to write more in each dataclass than is required
as the input to the book's generate_ast tool).

We're also not constrained by the one class per-file limitation, and do
not (currently) need an Expr base class to represent the type of all
expressions.

We should be able to avoid needing to implement the Visitor Pattern in
5.3, too, so these expression dataclasses won't need to be
modified/thrown away shortly.