[5.4] Add ast printing.
We don't need to implement the full Visitor Pattern in Python, we can
get away with the (IMO) cleaner single dispatch. I believe this gives
the same level of static type checking we'd get with the Visitor
Pattern, but is more readable - module some hacks to workaround mypy's
lack of support for @overload on an @singledispatch function.