Work around mypy limitations for unary functions.
The builtin neg() requires that its argument be negatable; the 'object'
type is not, but e.g. 'float' is.

mypy (0.941) cannot infer that if _is_numeric(x) returns true, x must be
a float. Instead, we can guard the call to neg(x) directly with an
isinstance() check to keep mypy happy. We then do not need to repeat
this check earlier with a _unary_op_check function.

This implementation ought to be repeated for the binary ops, too; for
whatever reason, in that case mypy interprets the op as,
  def (Any, Any) -> Any
and so is perfectly happy when we call op(left, right).