Fix resolver and Environment for for loops.
Tokens corresponding to different tokens in the source code now have
distinct values, even where the token type, lexeme and line are all
identical. This allows correct resolution of variables in for loops.

Note that e.g.,

  for (var i = 0; i <= 10; i = i + 1) { ... }

is desugared to (something like),

  var i = 0;
  while (i <= 10) {
    { ... }
    i = i + 1;
  }

and so the resolved binding depth for Variable 'i' in 'i <= 10' is less
than the binding depth for Variable 'i' in 'i = i + 1'. But, since all
those occurrences appear on the same line, they had identical Token
values, hence their Variable(token) values were also identical, and so
they hashed to the same value. The Bindings dictionary therefore treated
the two variables as identical, and it was not possible to correctly
resolve both instances of Variable 'i' in the for loop to *different*
depth values.

With the previous implementation, the below *would* work as expected,

  for (var i = 0; i <= 10;
       i = i + 1) { ... }

Because now the conflicting Variable expressions for 'i' are on
different lines, and hash to different values.

This commit adds a new item to all Tokens identifying the column at
which they start, which apart from being potentially useful for
debugging, also ensures that all instances of the same Variable will
have unique values and hence be distinguishable even when used as the
keys of a dict.