Attempt to force dependabot to use Python 3.9.
As of now (Nov 2022), generating poetry.lock with Python 3.11 breaks it
for both 3.9 and 3.10.

One specific case is tomli getting dropped. It is typically not required
by packages for Python >= 3.11 (e.g., black) because 3.11 added TOML
support to the stdlib. But dropping it completely from poetry.lock then
breaks 3.9 and 3.10.

Dependabot checks the python version in pyproject.toml, but (I think) it
(correctly) deems 3.11 as compatible with the ^3.9 specifier, hence uses
3.11. It also checks .python-version and runtime.txt. The former is more
conventional, but I don't want to force every pyenv user into 3.9! So
trying the latter.

https://github.com/dependabot/dependabot-core/blob/c5e5b827af80f60482262ffca0bd9a2bb7a4d73d/python/lib/dependabot/python/file_parser/python_requirement_parser.rb#L80
https://github.com/dependabot/dependabot-core/blob/c5e5b827af80f60482262ffca0bd9a2bb7a4d73d/python/lib/dependabot/python/file_parser/python_requirement_parser.rb#L20-L28
https://github.com/dependabot/dependabot-core/blob/c5e5b827af80f60482262ffca0bd9a2bb7a4d73d/python/lib/dependabot/python/update_checker/poetry_version_resolver.rb#L225-L233