Fix tox invocations for tox4.
tox4 requires a -- to separate arguments to tox to those of the command
it runs. We always had this, but (at least for the Makefile invocation),
this seems to interfere with/be interfered with when tox is invoked with
poetry run, as poetry run also uses -- to separate arguments.

It seems that running via make is an issue specifically, as directly
shell invocations of lines from the Makefile do in fact work as
expected. That's probably why it was working fine in CI.

This also adds the 'run' command to tox invocations, rather than letting
it default to the legacy entrypoint.